pyproject.toml
[tool.poetry]
name = "google-search-seo"
version = "0.1.0"
description = ""
authors = ["teamlu <lutp44@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
serpapi = "^0.1.5"
google-search-results = "^2.4.2"
python-dotenv = "^1.0.1"
pandas = "^2.2.1"
fuzzywuzzy = "^0.18.0"
python-levenshtein = "^0.25.0"
scikit-learn = "^1.4.1.post1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
